+++ src/picasapy/app/i18n/picasapy_hu.ts
@@ -8930,6 +8930,11 @@ A Picasa által esetleg figyelmen kívül hagyott vörösszemeket manuálisan ki
         <translation>Képpontméret</translation>
     </message>
     <message>
+        <location filename="../qml/PicasaPy/EditorParamPanel.qml"/>
+        <source>Reverse</source>
+        <translation>Megfordítás</translation>
+    </message>
+    <message>
         <source>Brush Size</source>
         <translation>Ecsetméret</translation>
     </message>